% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident data for state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer representing state for which data required}

\item{year}{Year of data required  (e.g. 2013)}
}
\value{
This function returns a map of the given state with accidents plotted for the given year. If invalid state number provided, error occurs and message returned.
If no data exists for provided year and state_num parameters, "no accidents to plot" message trigged and NULL returned.
}
\description{
This function reads in a FARS .csv format data file for a given year, extracts data for a given state number, and plot accidents for that state on a map.
}
\examples{
fars_map_state(1, 2013)

}
