% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in .csv format FARS data file to a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the data file}
}
\value{
This function returns a tibble containing the FARS data. If invalid filename provided, returns error with message that file does not exist.
}
\description{
This function reads a in .csv format FARS data file to a data.frame.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
