% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multiple FARS .csv format data files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years (e.g. c(2013, 2014))}
}
\value{
This function returns a tibble with the counts of observations in each year/month
}
\description{
This function reads in multiple FARS .csv format data files and summarises the data by returning count observations in each year/month category.
}
\examples{
fars_summarize_years(c(2013, 2014))

}
