% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the data file in YYYY format (e.g. 2013)}
}
\value{
This function returns a string representing a valid FARS filename format
}
\description{
This function creates a FARS filename string from a provided year input.
}
\examples{
make_filename("2013")

}
